qepcad-B.1.54
15/04/2010:Modified code to compile properly for 64-bit architectures.
           This relies on having at least saclib2.2.3 and tests the
           __WORDSIZE macro to see whether its value is 32 or 64.

+++ qesource/source/convenientstreams.h
@@ -10,9 +10,6 @@
  ***************************************************************/
 #ifndef _CONVENIENT_STREAMS_
 #define _CONVENIENT_STREAMS_
-#include <cstring>
-#include <cstdlib>
-#include <cstdio>
 #include <string>
 #include <sstream>
 using namespace std;
@@ -82,7 +79,7 @@ cacInBuff::int_type cacInBuff::underflow()
 {
   if (gptr() >= egptr())
   {
-    int leftover = min((long int)extra, (long int)(gptr() - eback()));
+    int leftover = min(extra, gptr() - eback());
     memmove(buff + (extra-leftover), gptr() - leftover, leftover);
 
     char *s = buff + extra;     // the string I'm reading in
